Implement deterministic nonce generation.
Since we want deterministic encryption, we need deterministic nonces.

Deterministic nonces are safe to use despite the admonition "never reuse nonces" because it's not re-use, per se, and certainly not dangerous reuse, when the entire cleartext is *also* the same: this simply results in identical *ciphertext*, which is absolutely no additional useful information to a hypothetical attacker.

We end up using Keccak here (the engine of SHA3) to construct the deterministic nonce.  See the documentation inside the relevant method for why we did this in addition to the nacl primitives.

Signed-off-by: Eric Myhre <hash@exultant.us>